[Chore] Challenge 3 in own branch
