+++ app.py
@@ -1,4 +1,4 @@
 from versions import app
 
 if __name__ == '__main__':
-    app.run()
+    app.run(host='0.0.0.0', port=8080)
