+++ versions/__init__.py
@@ -12,8 +12,7 @@ and its an easy read
 import os
 import jwt
 from functools import wraps
-from flask import Flask, request, jsonify, render_template, session
-from versions.v1.models import User, Business, Review
+from flask import Flask, request, jsonify
 from flask_cors import CORS
 from flask_mail import Mail
 from flask_sqlalchemy import SQLAlchemy
@@ -26,12 +25,6 @@ CORS(app)
 mail = Mail(app)
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
-
-
-user_instance = User()
-business_instance = Business()
-review_instance = Review()
-
 
 def login_required(f):
     """Ensures user is logged in before action
@@ -53,7 +46,7 @@ def login_required(f):
 
         is_token_valid = is_token_valid.valid if is_token_valid else True
 
-        if token in user_instance.tokens and user_instance.tokens[token] == 'invalid' or not is_token_valid:
+        if not is_token_valid:
             return jsonify({ 'warning': 'Login again'}), 401
 
         try:
@@ -71,43 +64,13 @@ def login_required(f):
         return f(current_user, *args, **kwargs)
     return wrap
 
-
-@app.route('/version1')
-def version1():
-    """route for API documentation"""
-    return render_template('version1.html')
-
-
-@app.route('/')
-def version2():
-    """route for API documentation"""
-    return render_template('version2.html')
-    
-@app.errorhandler(404)
-def page_not_found(e):
-    return jsonify({'warning': '404, Endpoint not found'}), 404
-
-@app.errorhandler(500)
-def internal_server_error(e):
-    return jsonify({'warning': '500, Internal Server Error'}), 500
-
-
-import versions.v1.user
-import versions.v1.get_user
-import versions.v1.business
-import versions.v1.review
+import versions.routes
 import versions.v2.models
 import versions.v2.auth
 import versions.v2.user
 import versions.v2.business
 import versions.v2.review
 import versions.v2.notifications
-
-# version 1 routes
-app.register_blueprint(versions.v1.user.mod, url_prefix='/api/v1/auth')
-app.register_blueprint(versions.v1.get_user.mod, url_prefix='/api/v1')
-app.register_blueprint(versions.v1.business.mod, url_prefix='/api/v1')
-app.register_blueprint(versions.v1.review.mod, url_prefix='/api/v1')
 
 # version 2 routes
 app.register_blueprint(versions.v2.auth.mod, url_prefix='/api/v2/auth')
@@ -116,4 +79,4 @@ app.register_blueprint(
     versions.v2.business.mod, url_prefix='/api/v2/businesses')
 app.register_blueprint(
     versions.v2.review.mod, url_prefix='/api/v2/businesses')
-app.register_blueprint(versions.v2.notifications.mod, url_prefix='/api/v2')
+app.register_blueprint(versions.v2.notifications.mod, url_prefix='/api/v2/notifications')
