+++ versions/templates/email.html
@@ -650,7 +650,7 @@
                 </tr>
                 <tr>
                   <td align="center" style="background-color:#f0f0f0" bgcolor="#f0f0f0">
-                    <a href="{{ url_for('business.read_all_businesses') }}" style="border:0 none; color:#5070e3; font-size: 55px; text-decoration:none" target="_blank">WeConnect
+                    <a href="{{ url_for('business_v2.read_all_businesses') }}" style="border:0 none; color:#5070e3; font-size: 55px; text-decoration:none" target="_blank">WeConnect
                     </a>
                   </td>
                 </tr>
